# apps/api/docs/swagger.yaml
definitions:
...
  cashflow.IgnoreTransactionsByFilterRequest:
    properties:
      filters:
        $ref: '#/definitions/transport_http_handlers_cashflow.TransactionFilters'
      ignored:
        type: boolean
    type: object
  cashflow.IgnoreTransactionsBySelectionRequest:
    properties:
      ids:
        items:
          type: string
        type: array
      ignored:
        type: boolean
    type: object
  cashflow.IgnoreTransactionsResponse:
    properties:
      status:
        type: string
      updated_count:
        type: integer
    type: object
...
  portfolio.RebuildPortfolioRequest:
    type: object
  portfolio.SnapshotPointResponse:
    properties:
      daily_return_pct:
        type: number
      market_value:
        type: integer
      occurred_at:
        type: string
      return_vs_cost_basis_pct:
        type: number
      time_weighted_return_pct:
        type: number
      total_cost_basis:
        type: integer
      total_pnl:
        type: integer
      total_pnl_pct:
        type: number
      value_index:
        type: number
    type: object
  transport_http_handlers_cashflow.TransactionFilters:
    properties:
      description:
        type: string
      direction:
        type: string
      from:
        type: string
      hide_ignored:
        type: boolean
      note:
        type: string
      q:
        type: string
      source:
        type: string
      tags:
        type: string
      to:
        type: string
      untagged:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /cashflow/transactions/ignore/filter:
    post:
      consumes:
      - application/json
      description: Set ignored=true/false for all transactions that match the supplied
        filter.
      parameters:
      - description: Filter ignore request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cashflow.IgnoreTransactionsByFilterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cashflow.IgnoreTransactionsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ignore filtered cashflow transactions
      tags:
      - Transactions
  /cashflow/transactions/ignore/selection:
    post:
      consumes:
      - application/json
      description: Set ignored=true/false for selected transaction IDs.
      parameters:
      - description: Selection ignore request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cashflow.IgnoreTransactionsBySelectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cashflow.IgnoreTransactionsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ignore selected cashflow transactions
      tags:
      - Transactions
...
  /portfolio/rebuild:
    post:
      consumes:
      - application/json
      description: Rebuilds portfolio positions and snapshots for the selected account.
      parameters:
      - description: Rebuild request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/portfolio.RebuildPortfolioRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rebuild portfolio
      tags:
      - portfolio
  /portfolio/snapshots:
    get:
      consumes:
      - application/json
      description: Returns snapshot points for the given account and optional date
        range.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/portfolio.SnapshotPointResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get portfolio snapshots
      tags:
      - portfolio
...
swagger: "2.0"